backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.12
openai>=1.0.0
chromadb>=0.4.0
langchain-text-splitters>=0.3.0
tiktoken>=0.7.0
pymupdf>=1.24.0
sentence-transformers>=3.0.0
